bettr: a better way to explore what is best

The bettr package provides a better way to explore what is best :) Details about how to use the package can be found in the vignette. The main entry point is the bettr() function, which opens an interactive application for exploring data consisting of multiple parallel rankings of a set of entities (e.g., computational methods ranked by their performance based on several different metrics).

See also

Author

Charlotte Soneson charlottesoneson@gmail.com

Federico Marini marinif@uni-mainz.de